Backend/requirements.txt
flask==3.0.3
flask_cors==5.0.0
mysql-connector-python==9.0.0